Fixed AlarmSystem exception when no GPIO pin set (#47)
Exception was seen by @ljsanchez #41 (comment)

When a GPIO pin isn't configured in the alarm config, AlarmSystem was not initializing the alarm status. When main.py attempted to check the alarm system's state it cause an exception. Instead we should always initialize the alarm stauts in AlarmSystem, even if a user is only going to monitor the system, and won't be arming/disarming (which is what the GPIO config is for.)